Forget Password & Reset Password feature is now complete.
Removed some unnecessary lines of code, and touched up on some parts of the code.

Added a "back to login" link to both the Forget Password and Reset Password pages for user to login in case they suddenly remember their password.

Added a success message upon success of resetting password.