Add Course button now saves to database
Turned off some fields in Task and Course, so that they will not be required for now. Remember to turn them on again next time the app is more developed.

As of now, the add course button saves the new course to the database, but fetching data from the database upon user login is still not done yet.