Add Task now updates and fetches from database.
Add Task can now update the database when a new task is added to the landing page.

Upon entering the landing page, the tasks belonging to the user are now fetched and displayed on the landing page.

Changed the checked component of the checkbox in the Tasks section to refer to the backend "status" attribute of task, rather than the frontend "completed" state.
task.status has 2 values "Todo" and "Done".