Updated the delete course and tasks functions
Initially, deleting the course and task only removes from the "course" and "task" section of the database. The user's  individual "courses" and "tasks" array were not updated, causing an unnecessary build up of course and task ids that no longer exists. For example, a user could have 30 objects in their tasks array, but in fact only 10 of them exists, as the rest were already deleted, just not from the array.

Updated the function such that the user's arrays are updated accordingly as well.