Minor changes to login page
Changed the name requirements from 3 characters to just a required field for sign up page.

Changed the error message for login page from "Email or Password does not exist." to "Incorrect Password." as this error message is more specific, and we already have an error message called "User not found." when the email exists in the database.