Half written code for integration testing of routes
My home internet connection went out, so I am trying to commit my changes such that I can save and access it on my laptop somewhere else.

The code might not work, as I was unable to test it out. If the code cannot run, please use the next latest version of the code.