Changed the checking of a strong password to be consistent with that of "Sign Up"
Initially, we used the front-end to check for whether the provided password is a strong password, i.e. minimum of 8 characters, with 1 uppercase, 1 lowercase character, and 1 number.

The "Sign Up" function checks the validity of the strong password using express.js, so I changed the checking of strong password of "Reset Password" to be consistent with that of the "Sign Up" function.

Now, both functions uses express.js to check if the password provided is a strong password before proceeding to save the user in the database.