Modified the code to only pass the user's id, course's id from frontend to the backend.
While creating integration testing for the course route, I encountered a 413 Payload Too Large error. This is possibly caused due to the initial passing of the whole user object to create a new course. Instead, I now changed the code to only pass the userid from the frontend to the backend while creating a new course, a new task, etc... This should solve the 413 Payload Too Large error.