Fixed an issue where the task was not associated to any user or course.
In the latest commit, the task and user id was passed in the request to the backend, instead of the actual task and user object itself. This caused a bug where the new task is now created with out a user nor course object, as the task id and user id are not Task nor Course.

This issue is now fixed, and the function works as normal.