Updated Integration Testing
Previously, the test cases of different files cannot run concurrently, as the database is being wiped and created concurrently with each file. This causes the state of the database to be mutated halfway through the testing and causes errors in testing.

A solution to this is to create 3 separate databases for route testing as they cannot be tested concurrently as they are dependent on the state of the database.

However, there may be more efficient solutions that I have not found.