New sort function on the landing page
Users can now sort their task list according to date or priority level.

Note that this feature is not yet complete as we have yet to implement the individual course priority levels.

As of now, sorting is not affected by course priority levels.

Also modified the User model to have two new attributes "tasksByDate" and "tasksByPriority" which are arrays of tasks sorted by date, and arrays of tasks sorted by priority respectively, instead of the original "tasks" array.