Notifications
#16
As of now, we have a mock notification that informs the user that their task has been created 5 seconds ago.

In the future, we would want the notification to inform users of their deadlines.

The current notification is just a placeholder to show that the notification is correctly shown at the correct timing.